+++ src/spm_redistribute.c
@@ -11,7 +11,8 @@
  * @author Pierre Ramet
  * @author Mathieu Faverge
  * @author Tony Delarue
- * @date 2023-01-11
+ * @author Alycia Lisito
+ * @date 2023-12-06
  *
  * @ingroup spm_dev_mpi
  * @{
@@ -163,7 +164,7 @@ spm_redist_get_newg2l( const spmatrix_t *oldspm,
     newspm.gN       = oldspm->gN;
     newspm.loc2glob = (spm_int_t *)newl2g;
     newspm.glob2loc = NULL;
-    spm_get_glob2loc( &newspm );
+    spm_getandset_glob2loc( &newspm );
 
     return newspm.glob2loc;
 }
@@ -247,7 +248,9 @@ spm_redist_extract_local( const spmatrix_t *oldspm,
     newspm->fmttype = SpmIJV;
     newspm->colptr  = malloc( oldspm->nnz * sizeof( spm_int_t ) );
     newspm->rowptr  = malloc( oldspm->nnz * sizeof( spm_int_t ) );
-    newspm->values  = malloc( oldspm->nnzexp * fltsize );
+    if ( fltsize > 0 ) {
+        newspm->values  = malloc( oldspm->nnzexp * fltsize );
+    }
 
     /* Get the correct pointers according to the column/row distribution */
     if ( distribution & SpmDistByColumn ) {
@@ -305,8 +308,10 @@ spm_redist_extract_local( const spmatrix_t *oldspm,
                 newrow += nbrow;
 
                 /* Copy the value array */
-                memcpy( newval, oldval, valsize );
-                newval += valsize;
+                if ( fltsize > 0 ) {
+                    memcpy( newval, oldval, valsize );
+                    newval += valsize;
+                }
             }
             else {
                 owner = -newg2l[jg] - 1;
@@ -370,7 +375,9 @@ spm_redist_extract_local( const spmatrix_t *oldspm,
         }
         newspm->colptr = realloc( newspm->colptr, newspm->nnz * sizeof( spm_int_t ) );
         newspm->rowptr = realloc( newspm->rowptr, newspm->nnz * sizeof( spm_int_t ) );
-        newspm->values = realloc( newspm->values, newspm->nnzexp * fltsize );
+        if ( fltsize > 0 ) {
+            newspm->values = realloc( newspm->values, newspm->nnzexp * fltsize );
+        }
     }
 
     return;