+++ tests/spm_dist_genrhs_tests.c
@@ -11,7 +11,7 @@
  * @version 1.2.1
  * @author Mathieu Faverge
  * @author Tony Delarue
- * @date 2022-02-22
+ * @date 2023-12-06
  *
  **/
 #include <stdint.h>
@@ -27,15 +27,15 @@ char *typename[] = { "SpmRhsOne", "SpmRhsI", "SpmRhsRndX", "SpmRhsRndB" };
 
 int main (int argc, char **argv)
 {
-    spmatrix_t    original, spmd;
+    spmatrix_t    original, *spmd;
     spm_int_t     ldx;
     int           clustnum = 0;
     int           baseval, root = -1;
     int           rc, err = 0;
     int           dofidx;
     int           nrhs = 3;
-    size_t        sizeloc, sizedst;
-    void         *bloc, *bdst;
+    size_t        sizeglob, sizedist;
+    void         *bglob, *bdist;
     spm_rhstype_t type;
 
     MPI_Init( &argc, &argv );
@@ -43,11 +43,16 @@ int main (int argc, char **argv)
      * Get options from command line
      */
     rc = spmTestGetSpm( &original, argc, argv );
-
     if ( rc != SPM_SUCCESS ) {
         fprintf(stderr, "ERROR: Could not read the file, stop the test !!!\n");
         return EXIT_FAILURE;
     }
+
+    /*
+     * Gather if input is a distributed matrix for protection,
+     * CI should not used distributed matrices for this one
+     */
+    spmGatherInPlace( &original );
 
     if ( original.dof == 1 ) {
         dofidx = 0;
@@ -68,73 +73,81 @@ int main (int argc, char **argv)
     MPI_Comm_rank( MPI_COMM_WORLD, &clustnum );
 
     printf(" -- SPM GenRHS Test --\n");
-    sizeloc = spm_size_of( original.flttype ) * original.nexp * nrhs;
-    bloc    = malloc( sizeloc );
+    sizeglob = spm_size_of( original.flttype ) * original.nexp * nrhs;
+    bglob    = malloc( sizeglob );
 
     for( type = SpmRhsOne; type <= SpmRhsRndB; type++ )
     {
-        memset( bloc, 0xab, sizeloc );
+        memset( bglob, 0xab, sizeglob );
         ldx = spm_imax( 1, original.nexp );
 
         if ( spmGenRHS( type, nrhs, &original,
-                        NULL, ldx, bloc, ldx ) != SPM_SUCCESS ) {
-            fprintf( stderr, "Issue to generate the local rhs\n" );
+                        NULL, ldx, bglob, ldx ) != SPM_SUCCESS ) {
+            fprintf( stderr, "Issue to generate the global rhs\n" );
             continue;
         }
 
-        rc = spmScatter( &spmd, -1, &original, -1, NULL, 1, MPI_COMM_WORLD );
+        spmd = malloc( sizeof(spmatrix_t) );
+        rc = spmScatter( spmd, -1, &original, -1, NULL, 1, MPI_COMM_WORLD );
         if ( rc != SPM_SUCCESS ) {
             fprintf( stderr, "Failed to scatter the spm\n" );
             err++;
             continue;
         }
-
-        sizedst = spm_size_of( spmd.flttype ) * spmd.nexp * nrhs;
-        bdst    = malloc( sizedst );
+        sizedist = spm_size_of( spmd->flttype ) * spmd->nexp * nrhs;
+        bdist    = malloc( sizedist );
 
         for( baseval=0; baseval<2; baseval++ )
         {
-            spmBase( &spmd, baseval );
+            spmBase( spmd, baseval );
 
             if ( clustnum == 0 ) {
                 printf( " Case: %s - base(%d) - dof(%s) - root(%d) - type(%s): ",
-                        fltnames[spmd.flttype],
+                        fltnames[spmd->flttype],
                         baseval, dofnames[dofidx], root, typename[type] );
             }
 
-            memset( bdst, 0xab, sizedst );
-            ldx = spm_imax( 1, spmd.nexp );
-            if ( spmGenRHS( type, nrhs, &spmd,
-                            NULL, ldx, bdst, ldx ) != SPM_SUCCESS ) {
+            memset( bdist, 0xab, sizedist );
+            ldx = spm_imax( 1, spmd->nexp );
+            if ( spmd->fmttype == SpmIJV ) {
+                /* Initialize the field in advance to avoid multiple computations */
+                spm_getandset_glob2loc( spmd );
+            }
+
+            if ( spmGenRHS( type, nrhs, spmd,
+                            NULL, ldx, bdist, ldx ) != SPM_SUCCESS ) {
                 err++;
                 continue;
             }
 
-            switch( spmd.flttype ){
+            switch( spmd->flttype ){
             case SpmComplex64:
-                rc = z_spm_dist_genrhs_check( &spmd, nrhs, bloc, bdst, root );
+                rc = z_spm_dist_genrhs_check( spmd, type, nrhs, bglob, bdist );
                 break;
 
             case SpmComplex32:
-                rc = c_spm_dist_genrhs_check( &spmd, nrhs, bloc, bdst, root );
+                rc = c_spm_dist_genrhs_check( spmd, type, nrhs, bglob, bdist );
                 break;
 
             case SpmFloat:
-                rc = s_spm_dist_genrhs_check( &spmd, nrhs, bloc, bdst, root );
+                rc = s_spm_dist_genrhs_check( spmd, type, nrhs, bglob, bdist );
                 break;
 
             case SpmDouble:
             default:
-                rc = d_spm_dist_genrhs_check( &spmd, nrhs, bloc, bdst, root );
+                rc = d_spm_dist_genrhs_check( spmd, type, nrhs, bglob, bdist );
             }
             PRINT_RES(rc)
-        }
+                }
 
-        free( bdst );
-        spmExit( &spmd );
+        free( bdist );
+        if ( original.loc2glob == NULL ) {
+            spmExit( spmd );
+            free( spmd );
+        }
     }
 
-    free( bloc );
+    free( bglob );
     spmExit( &original );
     MPI_Finalize();