+++ tests/spm_redistribute_tests.c
@@ -10,7 +10,7 @@
  * @version 1.2.1
  * @author Tony Delarue
  * @author Mathieu Faverge
- * @date 2022-02-22
+ * @date 2023-12-06
  *
  **/
 #include <stdint.h>
@@ -164,11 +164,16 @@ main( int argc, char **argv )
      * Get options from command line
      */
     rc = spmTestGetSpm( &original, argc, argv );
-
     if ( rc != SPM_SUCCESS ) {
         fprintf( stderr, "ERROR: Could not read the file, stop the test !!!\n" );
         return EXIT_FAILURE;
     }
+
+    /**
+     * Gather if input is a distributed matrix for protection,
+     * CI should not used distributed matrices for this one
+     */
+    spmGatherInPlace( &original );
 
     spmPrintInfo( &original, stdout );