+++ tests/spm_tests.h
@@ -10,7 +10,7 @@
  * @version 1.2.1
  * @author Mathieu Faverge
  * @author Tony Delarue
- * @date 2022-02-22
+ * @date 2023-12-06
  *
  **/
 #ifndef _spm_tests_h_
@@ -49,14 +49,17 @@ typedef enum spm_l2gtype_e {
     SpmRandom
 } spm_l2gtype_t;
 
-typedef struct spm_doftype_s {
-    char type;
-    int  dofmax;
-}spm_doftype_t;
+typedef struct spm_test_s {
+    char         *filename;
+    spm_driver_t  driver;
+    char          doftype;
+    int           dofmax;
+    int           spmdist;
+} spm_test_t;
 
-void spmGetOptions( int argc,             char **argv,
-                    spm_driver_t *driver, char **filename,
-                    spm_doftype_t *doftype );
+void spmGetOptions( int         argc,
+                    char      **argv,
+                    spm_test_t *options );
 int  spmTestCompare( const spmatrix_t *spm1, const spmatrix_t *spm2 );
 
 void core_zplrnt( int m, int n, spm_complex64_t *A, int lda,
@@ -111,8 +114,8 @@ void z_spm_print_check( char *filename, const spmatrix_t *spm );
 int  z_spm_matvec_check( spm_trans_t trans, const spmatrix_t *spm );
 int  z_spm_norm_check( const spmatrix_t *spm );
 int  z_spm_dist_norm_check( const spmatrix_t *spm, const spmatrix_t *spmdist );
-int  z_spm_dist_genrhs_check( const spmatrix_t *spm, spm_int_t nrhs,
-                              const spm_complex64_t *bloc, const spm_complex64_t *bdst, int root );
+int  z_spm_dist_genrhs_check( const spmatrix_t *spm, spm_rhstype_t type, spm_int_t nrhs,
+                              const spm_complex64_t *bglob, spm_complex64_t *bdist );
 int  z_spm_dist_matvec_check( spm_trans_t trans, const spmatrix_t *spm );
 int  z_spm_sort_check_values( const spmatrix_t *spm1, const spmatrix_t *spm2 );
 
@@ -120,8 +123,8 @@ void c_spm_print_check( char *filename, const spmatrix_t *spm );
 int  c_spm_matvec_check( spm_trans_t trans, const spmatrix_t *spm );
 int  c_spm_norm_check( const spmatrix_t *spm );
 int  c_spm_dist_norm_check( const spmatrix_t *spm, const spmatrix_t *spmdist );
-int  c_spm_dist_genrhs_check( const spmatrix_t *spm, spm_int_t nrhs,
-                              const spm_complex32_t *bloc, const spm_complex32_t *bdst, int root );
+int  c_spm_dist_genrhs_check( const spmatrix_t *spm, spm_rhstype_t type, spm_int_t nrhs,
+                              const spm_complex32_t *bglob, spm_complex32_t *bdist );
 int  c_spm_dist_matvec_check( spm_trans_t trans, const spmatrix_t *spm );
 int  c_spm_sort_check_values( const spmatrix_t *spm1, const spmatrix_t *spm2 );
 
@@ -129,8 +132,8 @@ void d_spm_print_check( char *filename, const spmatrix_t *spm );
 int  d_spm_matvec_check( spm_trans_t trans, const spmatrix_t *spm );
 int  d_spm_norm_check( const spmatrix_t *spm );
 int  d_spm_dist_norm_check( const spmatrix_t *spm, const spmatrix_t *spmdist );
-int  d_spm_dist_genrhs_check( const spmatrix_t *spm, spm_int_t nrhs,
-                              const double *bloc, const double *bdst, int root );
+int  d_spm_dist_genrhs_check( const spmatrix_t *spm, spm_rhstype_t type, spm_int_t nrhs,
+                              const double *bglob, double *bdist );
 int  d_spm_dist_matvec_check( spm_trans_t trans, const spmatrix_t *spm );
 int  d_spm_sort_check_values( const spmatrix_t *spm1, const spmatrix_t *spm2 );
 
@@ -138,8 +141,8 @@ void s_spm_print_check( char *filename, const spmatrix_t *spm );
 int  s_spm_matvec_check( spm_trans_t trans, const spmatrix_t *spm );
 int  s_spm_norm_check( const spmatrix_t *spm );
 int  s_spm_dist_norm_check( const spmatrix_t *spm, const spmatrix_t *spmdist );
-int  s_spm_dist_genrhs_check( const spmatrix_t *spm, spm_int_t nrhs,
-                              const float *bloc, const float *bdst, int root );
+int  s_spm_dist_genrhs_check( const spmatrix_t *spm, spm_rhstype_t type, spm_int_t nrhs,
+                              const float *bglob, float *bdist );
 int  s_spm_dist_matvec_check( spm_trans_t trans, const spmatrix_t *spm );
 int  s_spm_sort_check_values( const spmatrix_t *spm1, const spmatrix_t *spm2 );
 
@@ -190,6 +193,7 @@ spm_norm_dist_print_result( double norms, double normd, double result, int clust
  */
 spm_int_t spm_create_loc2glob_continuous( const spmatrix_t *spm, spm_int_t **l2g_ptr );
 int       spm_get_distribution( const spmatrix_t *spm );
+spm_int_t *spm_getandset_glob2loc( spmatrix_t *spm );
 
 /**
  * spm_test_utils routine to factorize the tests